CV/05.ObjectDetection/requirements.txt
# before the install requirements you need to install numpy==1.23.5 & wheel
# pip install -r requirements 
# pip install --no-cache-dir -r requirements.txt
# i code everythink with VScode, Ubunty22.04 and python 3.10.12

cvzone==1.5.6
ultralytics==8.0.26
hydra-core==1.3.2
matplotlib==3.6.3
opencv-python==4.6.0.66
Pillow==9.4.0
PyYAML==6.0.1
requests==2.28.2
scipy==1.10.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1

# for sort.py
filterpy==1.4.5
scikit-image==0.19.3
lap==0.4.0

# OCR
pytesseract==0.3.10

pandas==2.0.2
easyocr==1.7.0